### 2019-01-16 #1180 @whedon
[PRE REVIEW]: stats19: A package for working with open road crash data
**Submitting author:** @robinlovelace (<a href="http://orcid.org/0000-0001-5679-6536">Robin Lovelace</a>)
**Repository:** <a href="https://github.com/ropensci/stats19" target ="_blank">https://github.com/ropensci/stats19</a>
**Version:** 0.1.1
