### 2019-01-16 #1180 @arfon
@Robinlovelace - could you link to the rOpenSci onboarding issue here please?
