### 2019-01-16 #1181 @arfon
Accepting because of https://github.com/ropensci/software-review/issues/266
