### 2019-01-16 #1181 @Robinlovelace
According to Zotero auto import it is (clearly this is wrong - will raise an issue):

```

@misc{noauthor_stats19_2019,
  title = {Stats19: {{A}} Package for Working with Open Road Crash Data},
  shorttitle = {Stats19},
  abstract = {Lovelace et al., (2019). stats19: A package for working with open road crash data. Journal of Open Source Software, 4(33), 1181, https://doi.org/10.21105/joss.01181},
  language = {en},
  journal = {The Journal of Open Source Software},
  howpublished = {http://joss.theoj.org},
  month = jan,
  year = {2019},
  doi = {10.21105/joss.01181}
}

```
