### 2019-01-16 #1181 @arfon
This should do it? https://www.doi2bib.org/bib/10.21105%2Fjoss.01181
